// src/Models/TasksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Tasks {
  ID?: number;
  Title?: string;
  "Projects#Id"?: number;
  Projects?: Record<string, unknown>;
  Description?: string;
  "AssignedTo#Id"?: number;
  AssignedTo?: Record<string, unknown>;
  StartDate?: string;
  EndDate?: string;
  "Priority#Id"?: number;
  Priority?: Record<string, unknown>;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  Author?: Record<string, unknown>;
  "Editor#Claims"?: string;
  Editor?: Record<string, unknown>;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: Record<string, unknown>;
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: Record<string, unknown>;
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: unknown;
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}